% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbercruncher.R
\name{numbercruncher}
\alias{numbercruncher}
\title{Crunches numbers}
\usage{
numbercruncher()
}
\value{
Nothing. No questions asked.
}
\description{
Crunches numbers
}
\author{
Maas van Steenbergen \email{mvansteenbergen@proton.me}
}
